niflib: fix some compiler warnings (contributed by guidoj, see issue #2788600)

+++ include/nif_math.h
@@ -18,7 +18,7 @@ namespace Niflib {
 #endif
 
 #ifndef _countof
-#define _countof(x) ((sizeof (x))/(sizeof((x)[0])))
+#define _countof(x) (int)((sizeof (x))/(sizeof((x)[0])))
 #endif
 
 //Forward declarations
